+++ src/test/testMergeSchemas.ts
@@ -24,6 +24,7 @@ import {
 } from './testingSchemas';
 import { forAwaitEach } from 'iterall';
 import { makeExecutableSchema } from '../schemaGenerator';
+import { IResolvers } from '../Interfaces';
 
 const testCombinations = [
   {
@@ -94,25 +95,29 @@ let enumTest = `
   }
 `;
 
-const enumSchema = makeExecutableSchema({
-  typeDefs: enumTest,
-  resolvers: {
-    Color: {
-      RED: '#EA3232',
-    },
-    NumericEnum: {
-      TEST: 1,
-    },
-    Query: {
-      color() {
-        return '#EA3232';
+let enumSchema: GraphQLSchema;
+
+if (process.env.GRAPHQL_VERSION !== '^0.11') {
+  enumSchema = makeExecutableSchema({
+    typeDefs: enumTest,
+    resolvers: {
+      Color: {
+        RED: '#EA3232',
       },
-      numericEnum() {
-        return 1;
+      NumericEnum: {
+        TEST: 1,
+      },
+      Query: {
+        color() {
+          return '#EA3232';
+        },
+        numericEnum() {
+          return 1;
+        },
       },
     },
-  },
-});
+  });
+}
 
 let linkSchema = `
   """
@@ -207,6 +212,26 @@ if (process.env.GRAPHQL_VERSION === '^0.11') {
       numericEnum: NumericEnum
     }
   `;
+
+  enumSchema = makeExecutableSchema({
+    typeDefs: enumTest,
+    resolvers: {
+      Color: {
+        RED: '#EA3232',
+      },
+      NumericEnum: {
+        TEST: 1,
+      },
+      Query: {
+        color() {
+          return '#EA3232';
+        },
+        numericEnum() {
+          return 1;
+        },
+      },
+    },
+  });
 
   linkSchema = `
     # A new type linking the Property type.
@@ -509,8 +534,27 @@ testCombinations.forEach(async combination => {
       });
 
       it('works with custom enums', async () => {
+        const localSchema = makeExecutableSchema({
+          typeDefs: enumTest,
+          resolvers: {
+            Color: {
+              RED: '#EA3232',
+            },
+            NumericEnum: {
+              TEST: 1,
+            },
+            Query: {
+              color() {
+                return '#EA3232';
+              },
+              numericEnum() {
+                return 1;
+              },
+            },
+          },
+        });
         const enumResult = await graphql(
-          enumSchema,
+          localSchema,
           `
             query {
               color
@@ -979,6 +1023,201 @@ bookingById(id: "b1") {
         const Booking = mergedSchema.getType('Booking') as GraphQLObjectType;
         expect(Booking.isTypeOf).to.equal(undefined);
       });
+
+      it('should merge resolvers when passed an array of resolver objects', async () => {
+        const Scalars = () => ({
+          TestScalar: new GraphQLScalarType({
+            name: 'TestScalar',
+            description: undefined,
+            serialize: value => value,
+            parseValue: value => value,
+            parseLiteral: () => null,
+          }),
+        });
+        const Enums = () => ({
+          NumericEnum: {
+            TEST: 1,
+          },
+          Color: {
+            RED: '#EA3232',
+          },
+        });
+        const PropertyResolvers: IResolvers = {
+          Property: {
+            bookings: {
+              fragment: 'fragment PropertyFragment on Property { id }',
+              resolve(parent, args, context, info) {
+                return info.mergeInfo.delegate(
+                  'query',
+                  'bookingsByPropertyId',
+                  {
+                    propertyId: parent.id,
+                    limit: args.limit ? args.limit : null,
+                  },
+                  context,
+                  info,
+                );
+              },
+            },
+          },
+        };
+        const LinkResolvers: (info: any) => IResolvers = info => ({
+          Booking: {
+            property: {
+              fragment: 'fragment BookingFragment on Booking { propertyId }',
+              resolve(parent, args, context) {
+                return info.mergeInfo.delegate(
+                  'query',
+                  'propertyById',
+                  {
+                    id: parent.propertyId,
+                  },
+                  context,
+                  info,
+                );
+              },
+            },
+          },
+        });
+        const Query1 = () => ({
+          Query: {
+            color() {
+              return '#EA3232';
+            },
+            numericEnum() {
+              return 1;
+            },
+          },
+        });
+        const Query2: (info: any) => IResolvers = () => ({
+          Query: {
+            delegateInterfaceTest(parent, args, context, info) {
+              return info.mergeInfo.delegate(
+                'query',
+                'interfaceTest',
+                {
+                  kind: 'ONE',
+                },
+                context,
+                info,
+              );
+            },
+            delegateArgumentTest(parent, args, context, info) {
+              return info.mergeInfo.delegate(
+                'query',
+                'propertyById',
+                {
+                  id: 'p1',
+                },
+                context,
+                info,
+              );
+            },
+            linkTest() {
+              return {
+                test: 'test',
+              };
+            },
+            node: {
+              // fragment doesn't work
+              fragment: 'fragment NodeFragment on Node { id }',
+              resolve(parent, args, context, info) {
+                if (args.id.startsWith('p')) {
+                  return info.mergeInfo.delegate(
+                    'query',
+                    'propertyById',
+                    args,
+                    context,
+                    info,
+                  );
+                } else if (args.id.startsWith('b')) {
+                  return info.mergeInfo.delegate(
+                    'query',
+                    'bookingById',
+                    args,
+                    context,
+                    info,
+                  );
+                } else if (args.id.startsWith('c')) {
+                  return info.mergeInfo.delegate(
+                    'query',
+                    'customerById',
+                    args,
+                    context,
+                    info,
+                  );
+                } else {
+                  throw new Error('invalid id');
+                }
+              },
+            },
+          },
+        });
+
+        const AsyncQuery: (info: any) => IResolvers = info => ({
+          Query: {
+            async nodes(parent, args, context) {
+              const bookings = await info.mergeInfo.delegate(
+                'query',
+                'bookings',
+                {},
+                context,
+                info,
+              );
+              const properties = await info.mergeInfo.delegate(
+                'query',
+                'properties',
+                {},
+                context,
+                info,
+              );
+              return [...bookings, ...properties];
+            },
+          },
+        });
+        const schema = mergeSchemas({
+          schemas: [
+            propertySchema,
+            bookingSchema,
+            productSchema,
+            scalarTest,
+            enumTest,
+            linkSchema,
+            loneExtend,
+            localSubscriptionSchema,
+          ],
+          resolvers: [
+            Scalars,
+            Enums,
+            PropertyResolvers,
+            LinkResolvers,
+            Query1,
+            Query2,
+            AsyncQuery,
+          ],
+        });
+
+        const mergedResult = await graphql(
+          schema,
+          `
+            query {
+              dateTimeTest
+              test1: jsonTest(input: { foo: "bar" })
+              test2: jsonTest(input: 5)
+              test3: jsonTest(input: "6")
+            }
+          `,
+        );
+        const expected = {
+          data: {
+            dateTimeTest: '1987-09-25T12:00:00',
+            test1: { foo: 'bar' },
+            test2: 5,
+            test3: '6',
+          },
+        };
+        expect(mergedResult).to.deep.equal(expected);
+      });
     });
 
     describe('fragments', () => {
@@ -1006,7 +1245,7 @@ fragment BookingFragment on Booking {
         const propertyResult = await graphql(
           propertySchema,
           `
-          ${propertyFragment}
+            ${propertyFragment}
             query {
               propertyById(id: "p1") {
                 ...PropertyFragment
@@ -1030,8 +1269,8 @@ fragment BookingFragment on Booking {
         const mergedResult = await graphql(
           mergedSchema,
           `
-          ${bookingFragment}
-          ${propertyFragment}
+            ${bookingFragment}
+            ${propertyFragment}
 
             query {
               propertyById(id: "p1") {
@@ -1129,6 +1368,197 @@ bookingById(id: "b1") {
             fragment PropertyFragment on Property {
               id
               name
+            }
+          `,
+        );
+
+        expect(mergedResult).to.deep.equal({
+          data: {
+            propertyById: {
+              id: 'p2',
+              name: 'Another great hotel',
+              bookings: [
+                {
+                  id: 'b4',
+                  customer: {
+                    name: 'Exampler Customer',
+                  },
+                  property: {
+                    id: 'p2',
+                    name: 'Another great hotel',
+                  },
+                },
+              ],
+            },
+          },
+        });
+      });
+
+      it('overlapping selections', async () => {
+        const propertyFragment1 = `
+fragment PropertyFragment1 on Property {
+  id
+  name
+  location {
+    name
+  }
+}
+    `;
+        const propertyFragment2 = `
+fragment PropertyFragment2 on Property {
+  id
+  name
+  location {
+    name
+  }
+}
+    `;
+        const bookingFragment = `
+fragment BookingFragment on Booking {
+  id
+  customer {
+    name
+  }
+  startTime
+  endTime
+}
+    `;
+
+        const propertyResult = await graphql(
+          propertySchema,
+          `
+            ${propertyFragment1}
+            ${propertyFragment2}
+            query {
+              propertyById(id: "p1") {
+                ...PropertyFragment1
+                ...PropertyFragment2
+              }
+            }
+          `,
+        );
+
+        const bookingResult = await graphql(
+          bookingSchema,
+          `
+            ${bookingFragment}
+            query {
+              bookingById(id: "b1") {
+                ...BookingFragment
+              }
+            }
+          `,
+        );
+
+        const mergedResult = await graphql(
+          mergedSchema,
+          `
+            ${bookingFragment}
+            ${propertyFragment1}
+            ${propertyFragment2}
+
+            query {
+              propertyById(id: "p1") {
+                ...PropertyFragment1
+                ...PropertyFragment2
+              }
+              bookingById(id: "b1") {
+                ...BookingFragment
+              }
+            }
+          `,
+        );
+
+        expect(mergedResult).to.deep.equal({
+          data: {
+            ...propertyResult.data,
+            ...bookingResult.data,
+          },
+        });
+      });
+
+      it('containing links and overlapping fragments on relation', async () => {
+        const mergedResult = await graphql(
+          mergedSchema,
+          `
+            query {
+              propertyById(id: "p2") {
+                id
+                ... on Property {
+                  name
+                  ...BookingFragment1
+                  ...BookingFragment2
+                }
+              }
+            }
+
+            fragment BookingFragment1 on Property {
+              bookings {
+                id
+                property {
+                  id
+                  name
+                }
+              }
+            }
+
+            fragment BookingFragment2 on Property {
+              bookings {
+                customer {
+                  name
+                }
+              }
+            }
+          `,
+        );
+
+        expect(mergedResult).to.deep.equal({
+          data: {
+            propertyById: {
+              id: 'p2',
+              name: 'Another great hotel',
+              bookings: [
+                {
+                  id: 'b4',
+                  customer: {
+                    name: 'Exampler Customer',
+                  },
+                  property: {
+                    id: 'p2',
+                    name: 'Another great hotel',
+                  },
+                },
+              ],
+            },
+          },
+        });
+      });
+
+      it('containing links and single fragment on relation', async () => {
+        const mergedResult = await graphql(
+          mergedSchema,
+          `
+            query {
+              propertyById(id: "p2") {
+                id
+                ... on Property {
+                  name
+                  ...BookingFragment
+                }
+              }
+            }
+
+            fragment BookingFragment on Property {
+              bookings {
+                id
+                customer {
+                  name
+                }
+                property {
+                  id
+                  name
+                }
+              }
             }
           `,
         );